extract token validation to util function
