+++ client/src/redux/user.ts
@@ -1,5 +1,5 @@
 import { createSlice } from "@reduxjs/toolkit";
-import jwt_decode, { JwtPayload } from "jwt-decode";
+import { validateTokens } from "utils/validate_tokens";
 
 interface UserState {
   user?: {
@@ -19,42 +19,7 @@ const initialState: UserState = {
   user: null,
 };
 
-// check if the auth token is expired, if so, request a new one
-if (bearerToken && refreshToken) {
-  const decodedBearer: JwtPayload = jwt_decode(bearerToken);
-  const decodedRefresh: JwtPayload = jwt_decode(refreshToken);
-  const currentTime = Date.now() / 1000;
-
-  if (decodedRefresh?.exp && decodedRefresh.exp < currentTime) {
-    // refresh token is expired, user needs to log in again
-    localStorage.removeItem("REFRESH_TOKEN");
-    localStorage.removeItem("BEARER_TOKEN");
-    localStorage.removeItem("USER");
-    await fetch("/api/auth/logout", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ refreshToken }),
-    });
-  } else if (decodedBearer?.exp && decodedBearer?.exp < currentTime) {
-    // auth token is expired, but refresh token is still valid, request a new auth token
-    const response = await fetch("/api/auth/refresh", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ refreshToken }),
-    });
-
-    if (response.ok) {
-      const { bearerToken } = await response.json();
-      localStorage.setItem("BEARER_TOKEN", bearerToken);
-    } else {
-      console.log("Error refreshing token");
-    }
-  }
-}
+validateTokens(bearerToken, refreshToken);
 
 if (userInfo && refreshToken && bearerToken) {
   initialState.user = {
